+++ lib/src/compiler/autotools.dart
@@ -52,7 +52,15 @@ class AutoToolsCompiler extends BaseCompiler {
     // check configure exists
     if (!File(join(sourceDir, 'configure')).existsSync()) {
       simpleLogger.i('configure not found');
-      await shell.run('autoreconf -i -v', workingDirectory: sourceDir);
+
+      // check autogen.sh exists
+      if (File(join(sourceDir, 'autogen.sh')).existsSync()) {
+        simpleLogger.i('autogen.sh found, run it.');
+        await shell.run('./autogen.sh', workingDirectory: sourceDir);
+      } else {
+        simpleLogger.i('autogen.sh not found, try run `autoreconf -i -v`.');
+        await shell.run('autoreconf -i -v', workingDirectory: sourceDir);
+      }
     }
 
     // check configure exists
@@ -115,6 +123,11 @@ class AutoToolsCompiler extends BaseCompiler {
     // cpu number
     final cpuNumber = envs.cpuCount;
     final opt = lib.options.joinWithSpace();
+
+    if (depPrefix.isEmpty) {
+      // ignore: parameter_assignments
+      depPrefix = installPrefix;
+    }
 
     final configureCmd =
         './configure --prefix=$depPrefix --exec-prefix $installPrefix --host $host $opt';